// apps/nuxt/src/types/color.ts
// !!!important!! New color added need to be added to the sass setting file `packages/web/src/assets/scss/setting/_color.scss`
export enum Color {
  blueFrance = 'blue-france',
  blueFranceLighted = 'blue-france--light',
  blueFranceLightnessed = 'blue-france--lightness',

  green = 'green',
  greenLighted = 'green--light',
  greenLightnessed = 'green--lightness',

  purple = 'purple',
  purpleLighted = 'purple--light',
  purpleLightnessed = 'purple--lightness',

  deepRed = 'deep-red',
  red = 'red',
  redLighted = 'red--light',
  redLightnessed = 'red--lightness',

  yellow = 'yellow',
  yellowLighted = 'yellow--light',
  yellowLightnessed = 'yellow--lightness',

  grey = 'grey',
  greyLighted = 'grey--light',
  greyLightnessed = 'grey--lightness',

  white = 'white',
  black = 'black'
}